+++ Script/data_migration.py
@@ -1,14 +1,13 @@
 from argparse import ArgumentParser, Namespace, RawTextHelpFormatter
 from distutils.version import StrictVersion
-from glob import glob
-from json import loads, dumps
 from logging import error, info, warning
 from pathlib import Path
+from subprocess import run
 from typing import Dict, NoReturn, Any, List, Tuple, Union
 
-from jsonpatch import JsonPatch, PatchOperation, InvalidJsonPatch
 from ruamel.yaml import YAML
 
+from _version import parse_version
 from util.logging import configure_logging
 
 
@@ -25,9 +24,8 @@ def _main() -> NoReturn:
     try:
         config: dict = YAML(typ="safe").load(Path(args.config))
         repository_config: Dict[str, str] = config["timeline_tracker_app_config"]["repositories_config"]
-        from _version import get_strict_version
-        app_version = get_strict_version()
-        _ensure_data_migrated_to_current_version(app_version, **repository_config)
+        from _version import APP_VERSION
+        _ensure_data_migrated_to_current_version(APP_VERSION, **repository_config)
     except Exception as e:
         error(f"Failure occurred during data migration: {e}", exc_info=e)
         exit(-1)
@@ -46,24 +44,24 @@ def _parse_arguments() -> Namespace:
 
 def _ensure_data_migrated_to_current_version(
         app_version: StrictVersion,
-        *, repository_type: str, json_repositories_directory_root: str = None) -> NoReturn:
-    if repository_type == "memory":
+        *, world_repo_class_path: str = None, json_repositories_directory_root: str = None, **_) -> NoReturn:
+    if world_repo_class_path == "adapter.persistence.in_memory_repositories.InMemoryWorldRepository":
         # No need to migrate data
         pass
-    elif repository_type == "json":
+    elif world_repo_class_path == "adapter.persistence.json_file_repositories.JsonFileWorldRepository":
         if json_repositories_directory_root is None:
             error("Config file specifies json type repositories but did not provide the directory root.")
             exit(-1)
         _ensure_json_data_migrated_to_current_version(app_version, json_repositories_directory_root)
     else:
-        error(f"Failed to check data repository: unhandled repository type '{repository_type}'")
+        error(f"Failed to check data repository: unhandled repository type '{world_repo_class_path}'")
         exit(-1)
 
 
 def _ensure_json_data_migrated_to_current_version(app_version: StrictVersion, json_repositories_directory_root: str) -> NoReturn:
     json_repository_path = Path(json_repositories_directory_root)
     data_version_file = json_repository_path.joinpath("repository_version.metadata")
-    data_version = StrictVersion(data_version_file.read_text(encoding="utf8") if data_version_file.exists() else "0.0")
+    data_version = parse_version(data_version_file.read_text(encoding="utf8") if data_version_file.exists() else "0.0")
     if data_version > app_version:
         error("Repository data version is ahead of the application, this is an invalid state. Aborting...")
         exit(-1)
@@ -77,81 +75,38 @@ def _ensure_json_data_migrated_to_current_version(app_version: StrictVersion, js
             error("Could not locate data migration files. (Was this script run from the Timeline Tracker API project root folder?)")
             exit(-1)
 
-        migration_instructions: List[Tuple[StrictVersion, str, Dict["str", Any]]] = sorted([
+        migration_instructions: List[Tuple[StrictVersion, Path]] = sorted([
             (
-                StrictVersion(migration_file.name.split("__")[0][1:]),
-                migration_file.name.split("__")[1].removesuffix(".json").replace("_", " "),
-                loads(migration_file.read_text(encoding="utf8"))
+                parse_version(migration_file.name.split("__")[0][1:]),
+                migration_file
             )
             for migration_file in json_file_repository_migrations_dir.iterdir()
-            if migration_file.name.endswith(".json")
+            if migration_file.name.endswith(".py")
         ])
-        for version, change_description, instructions in migration_instructions:
+        for version, migration_py_script in migration_instructions:
             if version <= data_version:
                 continue
             if version > app_version:
                 warning(f"A migration file exists for a version ({version}) greater than the application version {app_version}, ignoring.")
                 continue
-            info(f"Attempting to migrate data to version {version} ({change_description})")
-            _apply_json_data_migration(json_repository_path, **instructions)
-            json_repository_path.joinpath("repository_version.metadata").write_text(str(version), encoding="utf8")
-            info(f"Successfully migrated data to version {version}")
 
-        json_repository_path.joinpath("repository_version.metadata").write_text(str(app_version), encoding="utf8")
-        info(f"Successfully migrated data to version {app_version}")
+            _run_migration_script(json_repository_path, migration_py_script, version)
+            _update_data_version_file(json_repository_path, version)
+
+        _update_data_version_file(json_repository_path, app_version)
     exit(0)
 
 
-def _apply_json_data_migration(
-        json_repository_path: Path,
-        *, migrations: List[_JsonMigration] = None, warnings: List[str] = None, **kwargs
-) -> None:
-    valid_instructions = {"warnings", "migrations"}
-    if not isinstance(migrations, list):
-        error(f"Invalid json migration instructions. Must contain 'migrations' list.")
-        exit(-1)
-    if kwargs:
-        error(f"Invalid json migration instructions. Must be subset of {valid_instructions}, but was: {set(kwargs.keys())}")
+def _run_migration_script(json_repository_path: Path, migration_py_script: Path, version: StrictVersion) -> None:
+    migration_return_code = run(["python", migration_py_script.as_posix(), json_repository_path.as_posix()]).returncode
+    if migration_return_code != 0:
+        error(f"Failed to apply data migration to version {version}, non-zero return code (was {migration_return_code})")
         exit(-1)
 
-    if warnings:
-        for warning_message in warnings:
-            warning(warning_message)
-        if "y" != input("    Continue with migration? (y/N) ").lower():
-            info("Aborting as instructed by user.")
-            exit(-1)
-    migrated_data_by_path: Dict[str, Dict[str, Any]] = {}
-    for migration in migrations:
-        matching_files, json_patch = _load_json_migration(json_repository_path, **migration)
-        for file_path in matching_files:
-            if file_path.as_posix() in migrated_data_by_path:
-                json_data = migrated_data_by_path[file_path.as_posix()]
-                print(dumps(json_data, indent=2))
-            else:
-                json_data = loads(file_path.read_text(encoding="utf8"))
-            mutated = json_patch.apply(json_data)
-            migrated_data_by_path[file_path.as_posix()] = mutated
-    for file_path_raw, data in migrated_data_by_path.items():
-        Path(file_path_raw).write_text(dumps(data, indent=2), encoding="utf8")
 
-
-def _load_json_migration(json_repository_path: Path, path_matcher: str, data_patch: List[Dict[str, Any]]) -> Tuple[List[Path], JsonPatch]:
-    full_path_matcher = json_repository_path.joinpath(path_matcher.replace("<guid>", _UUID_MATCHER)).as_posix()
-    matching_files = [Path(path).resolve() for path in glob(full_path_matcher, recursive=True)]
-    json_patch = _validate_and_load_json_patch(data_patch)
-    return matching_files, json_patch
-
-
-def _validate_and_load_json_patch(raw_patch: List[Dict[str, Any]]) -> JsonPatch:
-    if not isinstance(raw_patch, list) or any([not isinstance(modification, dict) for modification in raw_patch]):
-        error(f"Failed to load modification instructions, was not a valid JSON Patch: '{raw_patch}'")
-        exit(-1)
-
-    try:
-        return JsonPatch([PatchOperation(operation).operation for operation in raw_patch])
-    except InvalidJsonPatch as e:
-        error(f"Failed to load modification instructions: {e}. Was '{raw_patch}'")
-        exit(-1)
+def _update_data_version_file(json_repository_path: Path, version: StrictVersion) -> None:
+    json_repository_path.joinpath("repository_version.metadata").write_text(str(version), encoding="utf8")
+    info(f"Data has been migrated to version {version}")
 
 
 if __name__ == "__main__":