+++ Source/Python/adapter/persistence/json_file_repositories.py
@@ -2,6 +2,7 @@ from json import dumps, loads
 from pathlib import Path
 from typing import Set, Type, Generic, TypeVar, Dict, Optional, List
 
+from _version import APP_VERSION, APP_VERSION_RAW, parse_version
 from application.requests.data_forms import JsonTranslator
 from domain.events import Event
 from domain.ids import PrefixedUUID, IdentifiedEntity
@@ -33,6 +34,14 @@ class _JsonFileIdentifiedEntityRepository(Generic[_T]):
             repo_path.mkdir()
         if not repo_path.is_dir():
             raise ValueError(f"The path '{repo_path}' is not a valid directory and cannot be used.")
+
+        # If the repository does not have a version associated with it, stamp it with current. If it has a version that is not the current
+        # version, reject it. External data migration is responsible for updating it.
+        metadata_version_path = root_repos_path.joinpath(_METADATA_VERSION_FILE)
+        if not metadata_version_path.exists():
+            metadata_version_path.write_text(APP_VERSION_RAW, "utf8")
+        elif parse_version(metadata_version_path.read_text("utf8")) != APP_VERSION:
+            raise ValueError(f"The path '{repo_path}' contains data associated with a different app version.")
 
         self._repo_path = repo_path
         self._entity_type = entity_type